format.js and fix source
