+++ packages/ruler/types/format.d.ts
@@ -1,0 +1,2 @@
+export function defaultLabelFormat(value: number, units?: import("@turf/helpers").Units): string;
+export function defaultLabelAreaFormat(valueSquareMeters: number, units?: import("@turf/helpers").AreaUnits): string;
